add read/write raw data/index message.
git-svn-id: http://code.taobao.org/svn/tfs@1898 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/tfs-tmp-trunk/src/message/dataserver_task_message.h
@@ -33,7 +33,7 @@ namespace tfs
         inline void set_source_id(const uint64_t source_id) {source_id_ = source_id;}
         inline uint64_t get_source_id() const {return source_id_;}
       private:
-        uint32_t block_id_;
+        uint64_t block_id_;
         uint64_t source_id_;
     };
